chore(utils): rename validation.py → prerequisites.py and hash.py → bcrypt.py
Two utils modules had names that didn't match their contents:

- dom/utils/validation.py held runtime prerequisite checks
  (validate_docker_available, validate_port_available,
   validate_config_file, validate_infrastructure_prerequisites,
   warn_if_privileged_port). These are environment/system checks, not
  value-validation rules. The latter live in dom/validation/. The test
  file was already named test_prerequisites.py, hinting at the right
  name.

- dom/utils/hash.py exposed a single bcrypt password helper, while
  dom/utils/hashing.py held the deterministic team-ID hashing utilities.
  The hash/hashing pair was easy to mis-import.

Rename to dom/utils/prerequisites.py and dom/utils/bcrypt.py and update
the four importers (InfraService, DockerClient, and the two test files
that reference the modules under their old names).

No behavior change; pure rename + import update.
